Create and implement useWaypointCalculations hook
